+++ src/utils/read-tsconfig.ts
@@ -2,79 +2,9 @@ import fs from 'fs';
 import path from 'path';
 import { parse } from 'jsonc-parser';
 import slash from 'slash';
-import { findUp } from './find-up';
 import type { TsConfigJson, TsConfigJsonResolved } from '../types';
 import { normalizePath } from './normalize-path';
-
-const pathExists = (filePath: string) => fs.existsSync(filePath);
-
-const safeJsonParse = (jsonString: string) => {
-	try {
-		return JSON.parse(jsonString);
-	} catch {}
-};
-
-function resolveExtends(
-	filePath: string,
-	directoryPath: string,	
-) {
-	if (filePath === '..') {
-		filePath += '/tsconfig.json';
-	}
-
-	// Relative path
-	if (filePath.startsWith('.')) {
-		let tsconfigPath = path.resolve(directoryPath, filePath);
-
-		if (
-			pathExists(tsconfigPath)
-			&& fs.statSync(tsconfigPath).isFile()
-		) {
-			return tsconfigPath;
-		}
-
-		if (!tsconfigPath.endsWith('.json')) {
-			tsconfigPath += '.json';
-
-			if (pathExists(tsconfigPath)) {
-				return tsconfigPath;
-			}
-		}
-	} else {
-		const nodeModulePath = path.join('node_modules', filePath);	
-		let currentPath = findUp(
-			directoryPath,
-			nodeModulePath,
-		);
-
-		if (currentPath) {
-			if (fs.statSync(currentPath).isDirectory()) {
-				const packageJsonpath = path.join(currentPath, 'package.json');
-	
-				if (pathExists(packageJsonpath)) {
-					const packageJsonContent = fs.readFileSync(packageJsonpath, 'utf8');
-					const packageJson = safeJsonParse(packageJsonContent);
-	
-					if (packageJson && 'tsconfig' in packageJson) {
-						currentPath = path.join(currentPath, packageJson.tsconfig);
-					} else {
-						currentPath = path.join(currentPath, 'tsconfig.json');
-					}
-				} else {
-					currentPath = path.join(currentPath, 'tsconfig.json');
-				}
-	
-				if (pathExists(currentPath)) {
-					return currentPath;
-				}
-			} else {
-				return currentPath;
-			}
-		}
-	}
-
-	throw new Error(`File '${filePath}' not found.`);
-}
+import { resolveExtends } from './resolve-extends';
 
 export function readTsconfig(
 	filePath: string,
